Revert "serial: stm32: retrieve data stuck in DMA FIFO"
There is a race condition while operating with the DMA on the serial port,
where there is an access to an invalid address (0xdead) leading to a
system crash on BT and Wi-Fi.
That invalid DMA access was introduced in commit 83771e186d2abeccc38b0a5e6841b068499622ca.
Revert that commit till a proper fix is found.

https://onedigi.atlassian.net/browse/DEL-9877

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>